requirements.txt
streamlit
pandas
streamlit-option-menu
mysql-connector-python
python-dotenv